+++ src/common/collections/tenants/server/publish.js
@@ -15,7 +15,7 @@ import { Tenants } from '../index.js';
  *  - managers: the list of ids of users which are allowed to managed this tenant using a scoped role
  *  - records: the list of validity records for this entity
  *  - closest: the closest record
- * Only publish the tenants the user is allowed to
+ * Only publishes the tenants the user is allowed to
  */
 Meteor.publish( TenantsManager.C.pub.tenantsAll.publish, async function(){
     const self = this;
@@ -28,14 +28,14 @@ Meteor.publish( TenantsManager.C.pub.tenantsAll.publish, async function(){
 
     const entitiesObserver = Entities.collection.find({}).observeAsync({
         added: async function( item ){
-            if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.access', userId, item._id )){
+            if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.read', userId, item._id )){
                 self.added( TenantsManager.C.pub.tenantsAll.collection, item._id, await Tenants.s.transformEntity( item ));
                 TenantsManager.s.eventEmitter.emit( 'added', item._id, await Tenants.s.transformEntity( item ));
             }
         },
         changed: async function( newItem, oldItem ){
             if( !initializing ){
-                if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.access', userId, newItem._id )){
+                if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.read', userId, newItem._id )){
                     self.changed( TenantsManager.C.pub.tenantsAll.collection, newItem._id, await Tenants.s.transformEntity( newItem ));
                     TenantsManager.s.eventEmitter.emit( 'changed', newItem._id, await Tenants.s.transformEntity( newItem ));
                 }
@@ -49,7 +49,7 @@ Meteor.publish( TenantsManager.C.pub.tenantsAll.publish, async function(){
 
     const recordsObserver = Records.collection.find({}).observeAsync({
         added: async function( item ){
-            if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.access', userId, item.entity )){
+            if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.read', userId, item.entity )){
                 Entities.collection.findOneAsync({ _id: item.entity }).then( async ( entity ) => {
                     if( entity ){
                         try {
@@ -69,7 +69,7 @@ Meteor.publish( TenantsManager.C.pub.tenantsAll.publish, async function(){
         },
         changed: async function( newItem, oldItem ){
             if( !initializing ){
-                if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.access', userId, newItem.entity )){
+                if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.read', userId, newItem.entity )){
                     Entities.collection.findOneAsync({ _id: newItem.entity }).then( async ( entity ) => {
                         if( entity ){
                             self.changed( TenantsManager.C.pub.tenantsAll.collection, entity._id, await Tenants.s.transformEntity( entity ));
@@ -129,7 +129,7 @@ Meteor.publish( TenantsManager.C.pub.closests.publish, async function(){
 
     // records are changed, added or removed for a given entity: have to recompute the closest
     const f_closestChanged = async function( entity_id ){
-        if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.access', userId, entity_id )){
+        if( await TenantsManager.isAllowed( 'pwix.tenants_manager.feat.read', userId, entity_id )){
             Records.collection.find({ entity: entity_id }).fetchAsync().then(( fetched ) => {
                 const closest = Validity.closestByRecords( fetched ).record;
                 const prev_closest = entities[entity_id];