+++ src/common/collections/entities/index.js
@@ -6,5 +6,3 @@ export { Entities } from './collection.js';
 
 import './fieldset.js';
 import './schema.js';
-
-//console.debug( 'defined Entities' );