+++ src/common/collections/records/schema.js
@@ -2,20 +2,23 @@
  * pwix:tenants-manager/src/common/collections/records/schema.js
  */
 
+import { Logger } from 'meteor/pwix:logger';
 import SimpleSchema from 'meteor/aldeed:simple-schema';
 import { Tracker } from 'meteor/tracker';
 
 import { Records } from './index.js';
 
+const logger = Logger.get();
+
 Tracker.autorun(() => {
     const fieldSet = Records.fieldSet?.get();
     if( Records.collectionReady.get() && fieldSet ){
         if( Records.collection.attachSchema ){
-            _verbose( TenantsManager.C.Verbose.ATTACHSCHEMA, '[INFO] attaching schema to Records.collection' );
+            logger.verbose({ verbosity: TenantsManager.configure().verbosity, against: TenantsManager.C.Verbose.ATTACHSCHEMA }, '[INFO] attaching schema to Records.collection' );
             Records.collection.attachSchema( new SimpleSchema( fieldSet.toSchema()), { replace: true });
             Records.collection.attachBehaviour( 'timestampable', { replace: true });
         } else {
-            _verbose( TenantsManager.C.Verbose.ATTACHSCHEMA, '[WARN] Records.collection.attachSchema is not a function' );
+            logger.verbose({ verbosity: TenantsManager.configure().verbosity, against: TenantsManager.C.Verbose.ATTACHSCHEMA }, '[WARN] Records.collection.attachSchema is not a function' );
         }
     }
 });