+++ src/common/js/index.js
@@ -6,7 +6,6 @@ import './global.js';
 import './constants.js';
 import './configure.js';
 import './i18n.js';
-import './trace.js';
 //
 import './collections-get.js';
 import './event-emitter.js';